+++ interview/models.py
@@ -163,6 +163,10 @@ class Candidate(models.Model):
 
         return res
 
+    @property
+    def name_slug(self):
+        return slugify(self.name)
+
     class Meta:
         verbose_name = _("Candidate")
 
@@ -325,7 +329,7 @@ class Process(models.Model):
     def get_absolute_url(self):
         from django.urls import reverse
 
-        return reverse("process-details", args=[str(self.id)])
+        return reverse("process-details", kwargs={"process_id": self.id, "slug_info": f"_{self.candidate.name_slug}"})
 
     def is_open(self):
         return self.state not in Process.CLOSED_STATE_VALUES
@@ -499,7 +503,13 @@ class Interview(models.Model):
     def get_absolute_url(self):
         from django.urls import reverse
 
-        return reverse("interview-minute", args=[str(self.id)])
+        return reverse(
+            "interview-minute",
+            kwargs={
+                "interview_id": self.id,
+                "slug_info": f"_{self.process.candidate.name_slug}-{self.interviewers_trigram_slug}-{self.rank}",
+            },
+        )
 
     @property
     def planning_request_sent(self):
@@ -529,6 +539,12 @@ class Interview(models.Model):
     def interviewers_str(self):
         if self.id:
             return ", ".join(i.user.get_full_name() for i in self.interviewers.all())
+        return ""
+
+    @property
+    def interviewers_trigram_slug(self):
+        if self.id:
+            return "-".join(i.user.trigramme for i in self.interviewers.all())
         return ""
 
     def trigger_notification(self):